feat: Add a new quick way to provide hard to discover dependencies
Listen for packages present in the work directory and add them to the
roadmap. The roadmap already try to avoid duplicate lines.

Warning: this feature will only add packages by alphabetical order,
which may not respect dependency cycle. In that case, you'll be obliged
to still use the ‘my.roadmap.sh’ file.